requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
billiard==3.6.4.0
black==22.10.0
celery==5.2.7
certifi==2022.5.18.1
cffi==1.15.0
channels==4.0.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
daphne==4.0.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.4
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-ckeditor==6.5.0
django-cors-headers==3.12.0
django-debug-toolbar==3.6.0
django-elasticsearch-dsl==7.2.2
django-js-asset==2.0.0
django-otp==1.1.3
django-otp-twilio==1.0.1
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
django-restframework==0.0.1
django-templated-mail==1.1.1
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
elastic-transport==8.1.2
elasticsearch==7.17.4
elasticsearch-dsl==7.4.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==22.10.0
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
pathspec==0.10.1
pep517==0.12.0
phonenumbers==8.12.49
Pillow==9.2.0
pip-tools==6.6.2
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pytest==7.1.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
redis==4.3.4
requests==2.27.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tomli==2.0.1
twilio==7.9.2
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
zope.interface==5.5.0
